requirements.txt
openai
streamlit
langchain-community
python-dotenv
langchain-experimental
langchain-openai
tabulate
pyzipcode
folium
streamlit-folium
faker
duckduckgo-search
langchainhub
langchain-core
langchain-text-splitters
validators
fastembed
docarray
chromadb